chore: bump ketchup lite (1.5.12)
