+++ package.json
@@ -14,7 +14,7 @@
   "devDependencies": {
     "@types/node": "^22.5.0",
     "fs-extra": "^11.2.0",
-    "ketchup-lite": "1.5.11",
+    "ketchup-lite": "1.5.12",
     "rimraf": "^6.0.0",
     "sass": "^1.77.8",
     "typescript": "5.5.4"
