+++ web/deploy/js/ketchup-lite/ketchup-lite.js
@@ -16,7 +16,7 @@ var patchBrowser = () => {
 
 patchBrowser().then(async (options) => {
   await globalScripts();
-  return bootstrapLazy([["kul-showcase_36",[[1,"kul-showcase-messenger"],[1,"kul-showcase-chat"],[1,"kul-showcase-compare"],[1,"kul-showcase",{"kulStyle":[1537,"kul-style"],"debugInfo":[32],"currentComponent":[32],"currentFramework":[32],"currentUtility":[32]}],[1,"kul-showcase-code"],[1,"kul-showcase-splash"],[1,"kul-showcase-upload"],[1,"kul-showcase-button"],[1,"kul-showcase-tabbar"],[1,"kul-showcase-toast"],[1,"kul-showcase-badge"],[1,"kul-showcase-image"],[1,"kul-showcase-tree"],[1,"kul-showcase-accordion"],[1,"kul-showcase-card"],[1,"kul-showcase-chart"],[1,"kul-showcase-chip"],[1,"kul-showcase-lazy"],[1,"kul-showcase-list"],[1,"kul-showcase-photoframe"],[1,"kul-showcase-progressbar"],[1,"kul-showcase-spinner"],[1,"kul-showcase-switch"],[1,"kul-showcase-textfield"],[1,"kul-showcase-article"],[1,"kul-showcase-drawer"],[1,"kul-showcase-header"],[1,"kul-showcase-kuldata"],[1,"kul-showcase-kuldates"],[1,"kul-showcase-kuldebug"],[1,"kul-showcase-kuldynamicposition"],[1,"kul-showcase-kullanguage"],[1,"kul-showcase-kulmanager"],[1,"kul-showcase-kulmath"],[1,"kul-showcase-kulscrollonhover"],[1,"kul-showcase-kultheme"]]],["kul-showcase-debug",[[1,"kul-showcase-debug",{"currentComponent":[32],"currentProps":[32],"invalidJson":[32]}]]],["kul-drawer",[[1,"kul-drawer",{"kulStyle":[1537,"kul-style"],"debugInfo":[32],"opened":[32],"close":[64],"getDebugInfo":[64],"getProps":[64],"isOpened":[64],"open":[64],"refresh":[64],"toggle":[64],"unmount":[64]}]]],["kul-header",[[1,"kul-header",{"kulStyle":[1,"kul-style"],"debugInfo":[32],"getDebugInfo":[64],"getProps":[64],"refresh":[64],"unmount":[64]}]]],["kul-messenger",[[1,"kul-messenger",{"kulAutosave":[1028,"kul-autosave"],"kulData":[1040],"kulStyle":[1,"kul-style"],"kulValue":[16],"debugInfo":[32],"chat":[32],"connectionStatus":[32],"covers":[32],"currentCharacter":[32],"editingStatus":[32],"history":[32],"hoveredCustomizationOption":[32],"saveInProgress":[32],"ui":[32],"getDebugInfo":[64],"getProps":[64],"refresh":[64],"reset":[64],"unmount":[64]}]]],["kul-compare",[[1,"kul-compare",{"kulData":[1040],"kulShape":[1537,"kul-shape"],"kulStyle":[1537,"kul-style"],"kulView":[1025,"kul-view"],"debugInfo":[32],"shapes":[32],"view":[32],"isLeftPanelOpened":[32],"isRightPanelOpened":[32],"leftShape":[32],"rightShape":[32],"getDebugInfo":[64],"getProps":[64],"refresh":[64],"unmount":[64]},null,{"kulData":["updateShapes"],"kulShape":["updateShapes"]}]]],["kul-upload",[[1,"kul-upload",{"kulLabel":[1537,"kul-label"],"kulRipple":[1540,"kul-ripple"],"kulStyle":[1537,"kul-style"],"kulValue":[8,"kul-value"],"debugInfo":[32],"selectedFiles":[32],"getDebugInfo":[64],"getProps":[64],"getValue":[64],"refresh":[64],"unmount":[64]}]]],["kul-toast",[[1,"kul-toast",{"kulCloseIcon":[1040],"kulCloseCallback":[16],"kulIcon":[1040],"kulTimer":[2,"kul-timer"],"kulMessage":[1537,"kul-message"],"kulStyle":[1537,"kul-style"],"debugInfo":[32],"getDebugInfo":[64],"getProps":[64],"refresh":[64],"unmount":[64]}]]],["kul-accordion",[[1,"kul-accordion",{"kulData":[1040],"kulRipple":[1540,"kul-ripple"],"kulStyle":[1537,"kul-style"],"debugInfo":[32],"expandedNodes":[32],"selectedNodes":[32],"getDebugInfo":[64],"getProps":[64],"getSelectedNodes":[64],"refresh":[64],"toggleNode":[64],"unmount":[64]}]]],["kul-chart",[[1,"kul-chart",{"kulAxis":[1025,"kul-axis"],"kulColors":[16],"kulData":[16],"kulLegend":[1,"kul-legend"],"kulSeries":[16],"kulSizeX":[1,"kul-size-x"],"kulSizeY":[1,"kul-size-y"],"kulStyle":[1,"kul-style"],"kulTypes":[16],"kulXAxis":[16],"kulYAxis":[16],"debugInfo":[32],"getDebugInfo":[64],"getProps":[64],"refresh":[64],"unmount":[64]}]]],["kul-lazy",[[1,"kul-lazy",{"kulComponentName":[1,"kul-component-name"],"kulComponentProps":[16],"kulRenderMode":[1,"kul-render-mode"],"kulShowPlaceholder":[4,"kul-show-placeholder"],"kulStyle":[1,"kul-style"],"debugInfo":[32],"isInViewport":[32],"getComponent":[64],"getDebugInfo":[64],"getProps":[64],"refresh":[64],"unmount":[64]}]]],["kul-photoframe",[[1,"kul-photoframe",{"kulPlaceholder":[16],"kulStyle":[1537,"kul-style"],"kulThreshold":[2,"kul-threshold"],"kulValue":[16],"debugInfo":[32],"isInViewport":[32],"getDebugInfo":[64],"getProps":[64],"refresh":[64],"unmount":[64]}]]],["kul-splash",[[1,"kul-splash",{"kulLabel":[1025,"kul-label"],"kulStyle":[1537,"kul-style"],"debugInfo":[32],"state":[32],"getDebugInfo":[64],"getProps":[64],"refresh":[64],"unmount":[64]}]]],["kul-switch",[[1,"kul-switch",{"kulDisabled":[1540,"kul-disabled"],"kulLabel":[1537,"kul-label"],"kulLeadingLabel":[1540,"kul-leading-label"],"kulRipple":[1540,"kul-ripple"],"kulStyle":[1537,"kul-style"],"kulValue":[4,"kul-value"],"debugInfo":[32],"value":[32],"getDebugInfo":[64],"getProps":[64],"getValue":[64],"refresh":[64],"setValue":[64],"unmount":[64]}]]],["kul-card",[[1,"kul-card",{"kulData":[1040],"kulLayout":[1537,"kul-layout"],"kulSizeX":[1537,"kul-size-x"],"kulSizeY":[1537,"kul-size-y"],"kulStyle":[1537,"kul-style"],"debugInfo":[32],"shapes":[32],"getDebugInfo":[64],"getProps":[64],"getShapes":[64],"refresh":[64],"unmount":[64]}]]],["kul-tree",[[1,"kul-tree",{"kulAccordionLayout":[1540,"kul-accordion-layout"],"kulData":[1040],"kulFilter":[1028,"kul-filter"],"kulInitialExpansionDepth":[1026,"kul-initial-expansion-depth"],"kulRipple":[1540,"kul-ripple"],"kulSelectable":[1540,"kul-selectable"],"kulStyle":[1537,"kul-style"],"debugInfo":[32],"expandedNodes":[32],"hiddenNodes":[32],"selectedNode":[32],"getDebugInfo":[64],"getProps":[64],"refresh":[64],"unmount":[64]}]]],["kul-code_2",[[1,"kul-code",{"kulFormat":[1540,"kul-format"],"kulLanguage":[1537,"kul-language"],"kulPreserveSpaces":[1540,"kul-preserve-spaces"],"kulStyle":[1537,"kul-style"],"kulValue":[1025,"kul-value"],"debugInfo":[32],"value":[32],"getDebugInfo":[64],"getProps":[64],"refresh":[64],"unmount":[64]}],[1,"kul-progressbar",{"kulCenteredLabel":[516,"kul-centered-label"],"kulIcon":[1,"kul-icon"],"kulIsRadial":[516,"kul-is-radial"],"kulLabel":[1,"kul-label"],"kulStyle":[1,"kul-style"],"kulValue":[2,"kul-value"],"debugInfo":[32],"getDebugInfo":[64],"getProps":[64],"refresh":[64],"unmount":[64]}]]],["kul-textfield",[[1,"kul-textfield",{"kulDisabled":[1540,"kul-disabled"],"kulFullWidth":[1540,"kul-full-width"],"kulHelper":[1040],"kulHtmlAttributes":[1040],"kulIcon":[1537,"kul-icon"],"kulLabel":[1537,"kul-label"],"kulStyle":[1537,"kul-style"],"kulStyling":[1537,"kul-styling"],"kulTrailingIcon":[1540,"kul-trailing-icon"],"kulValue":[1,"kul-value"],"debugInfo":[32],"status":[32],"value":[32],"getDebugInfo":[64],"getProps":[64],"getValue":[64],"refresh":[64],"setBlur":[64],"setFocus":[64],"setValue":[64],"unmount":[64]}]]],["kul-badge_3",[[1,"kul-badge",{"kulImageProps":[1040],"kulLabel":[1025,"kul-label"],"kulStyle":[1537,"kul-style"],"debugInfo":[32],"getDebugInfo":[64],"getProps":[64],"refresh":[64],"unmount":[64]}],[1,"kul-image",{"kulBadgeProps":[1040],"kulColor":[1537,"kul-color"],"kulShowSpinner":[1540,"kul-show-spinner"],"kulSizeX":[1537,"kul-size-x"],"kulSizeY":[1537,"kul-size-y"],"kulStyle":[1537,"kul-style"],"kulValue":[1537,"kul-value"],"debugInfo":[32],"getDebugInfo":[64],"getProps":[64],"refresh":[64],"unmount":[64]}],[1,"kul-spinner",{"kulActive":[1540,"kul-active"],"kulBarVariant":[1540,"kul-bar-variant"],"kulDimensions":[1537,"kul-dimensions"],"kulFader":[1540,"kul-fader"],"kulFaderTimeout":[1538,"kul-fader-timeout"],"kulFullScreen":[1540,"kul-full-screen"],"kulLayout":[1538,"kul-layout"],"kulStyle":[1537,"kul-style"],"debugInfo":[32],"getDebugInfo":[64],"getProps":[64],"refresh":[64],"unmount":[64]}]]],["kul-button_2",[[1,"kul-button",{"kulData":[1040],"kulDisabled":[1540,"kul-disabled"],"kulIcon":[1537,"kul-icon"],"kulIconOff":[1537,"kul-icon-off"],"kulLabel":[1537,"kul-label"],"kulRipple":[1540,"kul-ripple"],"kulShowSpinner":[1540,"kul-show-spinner"],"kulStyle":[1537,"kul-style"],"kulStyling":[1537,"kul-styling"],"kulToggable":[1540,"kul-toggable"],"kulTrailingIcon":[1540,"kul-trailing-icon"],"kulType":[1537,"kul-type"],"kulValue":[4,"kul-value"],"debugInfo":[32],"value":[32],"getDebugInfo":[64],"getProps":[64],"getValue":[64],"refresh":[64],"setValue":[64],"unmount":[64]}],[1,"kul-list",{"kulData":[1040],"kulEmptyLabel":[1,"kul-empty-label"],"kulEnableDeletions":[4,"kul-enable-deletions"],"kulNavigation":[4,"kul-navigation"],"kulRipple":[1540,"kul-ripple"],"kulSelectable":[4,"kul-selectable"],"kulStyle":[1025,"kul-style"],"debugInfo":[32],"focused":[32],"selected":[32],"focusNext":[64],"focusPrevious":[64],"getDebugInfo":[64],"getProps":[64],"getSelected":[64],"refresh":[64],"selectNode":[64],"unmount":[64]},[[0,"keydown","listenKeydown"]]]]],["kul-chat_3",[[1,"kul-chat",{"kulContextWindow":[1026,"kul-context-window"],"kulEndpointUrl":[1025,"kul-endpoint-url"],"kulLayout":[1537,"kul-layout"],"kulMaxTokens":[1026,"kul-max-tokens"],"kulPollingInterval":[2,"kul-polling-interval"],"kulSeed":[1026,"kul-seed"],"kulStyle":[1537,"kul-style"],"kulSystem":[1025,"kul-system"],"kulTemperature":[1026,"kul-temperature"],"kulValue":[1040],"debugInfo":[32],"history":[32],"status":[32],"toolbarMessage":[32],"view":[32],"getDebugInfo":[64],"getHistory":[64],"getLastMessage":[64],"getProps":[64],"refresh":[64],"setHistory":[64],"unmount":[64]},[[0,"keydown","listenKeydown"]],{"kulSystem":["updateTokensCount"]}],[1,"kul-tabbar",{"kulData":[16],"kulRipple":[1540,"kul-ripple"],"kulStyle":[1537,"kul-style"],"kulValue":[520,"kul-value"],"debugInfo":[32],"value":[32],"getDebugInfo":[64],"getProps":[64],"getValue":[64],"refresh":[64],"setValue":[64],"unmount":[64]}],[1,"kul-chip",{"kulData":[1040],"kulRipple":[1540,"kul-ripple"],"kulStyle":[1025,"kul-style"],"kulStyling":[1025,"kul-styling"],"debugInfo":[32],"expandedNodes":[32],"hiddenNodes":[32],"selectedNodes":[32],"getDebugInfo":[64],"getProps":[64],"getSelectedNodes":[64],"refresh":[64],"setSelectedNodes":[64],"unmount":[64]}]]],["kul-article",[[1,"kul-article",{"kulData":[1040],"kulStyle":[1537,"kul-style"],"debugInfo":[32],"getDebugInfo":[64],"getProps":[64],"refresh":[64],"unmount":[64]}]]]], options);
+  return bootstrapLazy(JSON.parse("[[\"kul-showcase_37\",[[1,\"kul-showcase-messenger\"],[1,\"kul-showcase-chat\"],[1,\"kul-showcase-compare\"],[1,\"kul-showcase\",{\"kulStyle\":[1537,\"kul-style\"],\"debugInfo\":[32],\"currentComponent\":[32],\"currentFramework\":[32],\"currentUtility\":[32]}],[1,\"kul-showcase-code\"],[1,\"kul-showcase-masonry\"],[1,\"kul-showcase-splash\"],[1,\"kul-showcase-upload\"],[1,\"kul-showcase-button\"],[1,\"kul-showcase-tabbar\"],[1,\"kul-showcase-toast\"],[1,\"kul-showcase-badge\"],[1,\"kul-showcase-image\"],[1,\"kul-showcase-tree\"],[1,\"kul-showcase-accordion\"],[1,\"kul-showcase-card\"],[1,\"kul-showcase-chart\"],[1,\"kul-showcase-chip\"],[1,\"kul-showcase-lazy\"],[1,\"kul-showcase-list\"],[1,\"kul-showcase-photoframe\"],[1,\"kul-showcase-progressbar\"],[1,\"kul-showcase-spinner\"],[1,\"kul-showcase-switch\"],[1,\"kul-showcase-textfield\"],[1,\"kul-showcase-article\"],[1,\"kul-showcase-drawer\"],[1,\"kul-showcase-header\"],[1,\"kul-showcase-kuldata\"],[1,\"kul-showcase-kuldates\"],[1,\"kul-showcase-kuldebug\"],[1,\"kul-showcase-kuldynamicposition\"],[1,\"kul-showcase-kullanguage\"],[1,\"kul-showcase-kulmanager\"],[1,\"kul-showcase-kulmath\"],[1,\"kul-showcase-kulscrollonhover\"],[1,\"kul-showcase-kultheme\"]]],[\"kul-showcase-debug\",[[1,\"kul-showcase-debug\",{\"currentComponent\":[32],\"currentProps\":[32],\"invalidJson\":[32]}]]],[\"kul-drawer\",[[1,\"kul-drawer\",{\"kulStyle\":[1537,\"kul-style\"],\"debugInfo\":[32],\"opened\":[32],\"close\":[64],\"getDebugInfo\":[64],\"getProps\":[64],\"isOpened\":[64],\"open\":[64],\"refresh\":[64],\"toggle\":[64],\"unmount\":[64]}]]],[\"kul-header\",[[1,\"kul-header\",{\"kulStyle\":[1,\"kul-style\"],\"debugInfo\":[32],\"getDebugInfo\":[64],\"getProps\":[64],\"refresh\":[64],\"unmount\":[64]}]]],[\"kul-messenger\",[[1,\"kul-messenger\",{\"kulAutosave\":[1028,\"kul-autosave\"],\"kulData\":[1040],\"kulStyle\":[1,\"kul-style\"],\"kulValue\":[16],\"debugInfo\":[32],\"chat\":[32],\"connectionStatus\":[32],\"covers\":[32],\"currentCharacter\":[32],\"editingStatus\":[32],\"history\":[32],\"hoveredCustomizationOption\":[32],\"saveInProgress\":[32],\"ui\":[32],\"getDebugInfo\":[64],\"getProps\":[64],\"refresh\":[64],\"reset\":[64],\"unmount\":[64]}]]],[\"kul-compare\",[[1,\"kul-compare\",{\"kulData\":[1040],\"kulShape\":[1537,\"kul-shape\"],\"kulStyle\":[1537,\"kul-style\"],\"kulView\":[1025,\"kul-view\"],\"debugInfo\":[32],\"shapes\":[32],\"view\":[32],\"isLeftPanelOpened\":[32],\"isRightPanelOpened\":[32],\"leftShape\":[32],\"rightShape\":[32],\"getDebugInfo\":[64],\"getProps\":[64],\"refresh\":[64],\"unmount\":[64]},null,{\"kulData\":[\"updateShapes\"],\"kulShape\":[\"updateShapes\"]}]]],[\"kul-masonry\",[[1,\"kul-masonry\",{\"kulColumns\":[1026,\"kul-columns\"],\"kulData\":[1040],\"kulSelectable\":[1540,\"kul-selectable\"],\"kulShape\":[1537,\"kul-shape\"],\"kulStyle\":[1537,\"kul-style\"],\"kulView\":[1025,\"kul-view\"],\"debugInfo\":[32],\"selectedShape\":[32],\"shapes\":[32],\"view\":[32],\"getDebugInfo\":[64],\"getProps\":[64],\"refresh\":[64],\"unmount\":[64]},null,{\"kulData\":[\"updateShapes\"],\"kulShape\":[\"updateShapes\"]}]]],[\"kul-upload\",[[1,\"kul-upload\",{\"kulLabel\":[1537,\"kul-label\"],\"kulRipple\":[1540,\"kul-ripple\"],\"kulStyle\":[1537,\"kul-style\"],\"kulValue\":[8,\"kul-value\"],\"debugInfo\":[32],\"selectedFiles\":[32],\"getDebugInfo\":[64],\"getProps\":[64],\"getValue\":[64],\"refresh\":[64],\"unmount\":[64]}]]],[\"kul-toast\",[[1,\"kul-toast\",{\"kulCloseIcon\":[1040],\"kulCloseCallback\":[16],\"kulIcon\":[1040],\"kulTimer\":[2,\"kul-timer\"],\"kulMessage\":[1537,\"kul-message\"],\"kulStyle\":[1537,\"kul-style\"],\"debugInfo\":[32],\"getDebugInfo\":[64],\"getProps\":[64],\"refresh\":[64],\"unmount\":[64]}]]],[\"kul-accordion\",[[1,\"kul-accordion\",{\"kulData\":[1040],\"kulRipple\":[1540,\"kul-ripple\"],\"kulStyle\":[1537,\"kul-style\"],\"debugInfo\":[32],\"expandedNodes\":[32],\"selectedNodes\":[32],\"getDebugInfo\":[64],\"getProps\":[64],\"getSelectedNodes\":[64],\"refresh\":[64],\"toggleNode\":[64],\"unmount\":[64]}]]],[\"kul-chart\",[[1,\"kul-chart\",{\"kulAxis\":[1025,\"kul-axis\"],\"kulColors\":[16],\"kulData\":[16],\"kulLegend\":[1,\"kul-legend\"],\"kulSeries\":[16],\"kulSizeX\":[1,\"kul-size-x\"],\"kulSizeY\":[1,\"kul-size-y\"],\"kulStyle\":[1,\"kul-style\"],\"kulTypes\":[16],\"kulXAxis\":[16],\"kulYAxis\":[16],\"debugInfo\":[32],\"getDebugInfo\":[64],\"getProps\":[64],\"refresh\":[64],\"unmount\":[64]}]]],[\"kul-lazy\",[[1,\"kul-lazy\",{\"kulComponentName\":[1,\"kul-component-name\"],\"kulComponentProps\":[16],\"kulRenderMode\":[1,\"kul-render-mode\"],\"kulShowPlaceholder\":[4,\"kul-show-placeholder\"],\"kulStyle\":[1,\"kul-style\"],\"debugInfo\":[32],\"isInViewport\":[32],\"getComponent\":[64],\"getDebugInfo\":[64],\"getProps\":[64],\"refresh\":[64],\"unmount\":[64]}]]],[\"kul-photoframe\",[[1,\"kul-photoframe\",{\"kulPlaceholder\":[16],\"kulStyle\":[1537,\"kul-style\"],\"kulThreshold\":[2,\"kul-threshold\"],\"kulValue\":[16],\"debugInfo\":[32],\"isInViewport\":[32],\"getDebugInfo\":[64],\"getProps\":[64],\"refresh\":[64],\"unmount\":[64]}]]],[\"kul-splash\",[[1,\"kul-splash\",{\"kulLabel\":[1025,\"kul-label\"],\"kulStyle\":[1537,\"kul-style\"],\"debugInfo\":[32],\"state\":[32],\"getDebugInfo\":[64],\"getProps\":[64],\"refresh\":[64],\"unmount\":[64]}]]],[\"kul-switch\",[[1,\"kul-switch\",{\"kulDisabled\":[1540,\"kul-disabled\"],\"kulLabel\":[1537,\"kul-label\"],\"kulLeadingLabel\":[1540,\"kul-leading-label\"],\"kulRipple\":[1540,\"kul-ripple\"],\"kulStyle\":[1537,\"kul-style\"],\"kulValue\":[4,\"kul-value\"],\"debugInfo\":[32],\"value\":[32],\"getDebugInfo\":[64],\"getProps\":[64],\"getValue\":[64],\"refresh\":[64],\"setValue\":[64],\"unmount\":[64]}]]],[\"kul-card\",[[1,\"kul-card\",{\"kulData\":[1040],\"kulLayout\":[1537,\"kul-layout\"],\"kulSizeX\":[1537,\"kul-size-x\"],\"kulSizeY\":[1537,\"kul-size-y\"],\"kulStyle\":[1537,\"kul-style\"],\"debugInfo\":[32],\"shapes\":[32],\"getDebugInfo\":[64],\"getProps\":[64],\"getShapes\":[64],\"refresh\":[64],\"unmount\":[64]}]]],[\"kul-tree\",[[1,\"kul-tree\",{\"kulAccordionLayout\":[1540,\"kul-accordion-layout\"],\"kulData\":[1040],\"kulFilter\":[1028,\"kul-filter\"],\"kulInitialExpansionDepth\":[1026,\"kul-initial-expansion-depth\"],\"kulRipple\":[1540,\"kul-ripple\"],\"kulSelectable\":[1540,\"kul-selectable\"],\"kulStyle\":[1537,\"kul-style\"],\"debugInfo\":[32],\"expandedNodes\":[32],\"hiddenNodes\":[32],\"selectedNode\":[32],\"getDebugInfo\":[64],\"getProps\":[64],\"refresh\":[64],\"unmount\":[64]}]]],[\"kul-code_2\",[[1,\"kul-code\",{\"kulFormat\":[1540,\"kul-format\"],\"kulLanguage\":[1537,\"kul-language\"],\"kulPreserveSpaces\":[1540,\"kul-preserve-spaces\"],\"kulStyle\":[1537,\"kul-style\"],\"kulValue\":[1025,\"kul-value\"],\"debugInfo\":[32],\"value\":[32],\"getDebugInfo\":[64],\"getProps\":[64],\"refresh\":[64],\"unmount\":[64]}],[1,\"kul-progressbar\",{\"kulCenteredLabel\":[516,\"kul-centered-label\"],\"kulIcon\":[1,\"kul-icon\"],\"kulIsRadial\":[516,\"kul-is-radial\"],\"kulLabel\":[1,\"kul-label\"],\"kulStyle\":[1,\"kul-style\"],\"kulValue\":[2,\"kul-value\"],\"debugInfo\":[32],\"getDebugInfo\":[64],\"getProps\":[64],\"refresh\":[64],\"unmount\":[64]}]]],[\"kul-textfield\",[[1,\"kul-textfield\",{\"kulDisabled\":[1540,\"kul-disabled\"],\"kulFullWidth\":[1540,\"kul-full-width\"],\"kulHelper\":[1040],\"kulHtmlAttributes\":[1040],\"kulIcon\":[1537,\"kul-icon\"],\"kulLabel\":[1537,\"kul-label\"],\"kulStyle\":[1537,\"kul-style\"],\"kulStyling\":[1537,\"kul-styling\"],\"kulTrailingIcon\":[1540,\"kul-trailing-icon\"],\"kulValue\":[1,\"kul-value\"],\"debugInfo\":[32],\"status\":[32],\"value\":[32],\"getDebugInfo\":[64],\"getProps\":[64],\"getValue\":[64],\"refresh\":[64],\"setBlur\":[64],\"setFocus\":[64],\"setValue\":[64],\"unmount\":[64]}]]],[\"kul-spinner\",[[1,\"kul-spinner\",{\"kulActive\":[1540,\"kul-active\"],\"kulBarVariant\":[1540,\"kul-bar-variant\"],\"kulDimensions\":[1537,\"kul-dimensions\"],\"kulFader\":[1540,\"kul-fader\"],\"kulFaderTimeout\":[1538,\"kul-fader-timeout\"],\"kulFullScreen\":[1540,\"kul-full-screen\"],\"kulLayout\":[1538,\"kul-layout\"],\"kulStyle\":[1537,\"kul-style\"],\"debugInfo\":[32],\"getDebugInfo\":[64],\"getProps\":[64],\"refresh\":[64],\"unmount\":[64]}]]],[\"kul-article\",[[1,\"kul-article\",{\"kulData\":[1040],\"kulStyle\":[1537,\"kul-style\"],\"debugInfo\":[32],\"getDebugInfo\":[64],\"getProps\":[64],\"refresh\":[64],\"unmount\":[64]}]]],[\"kul-badge_4\",[[1,\"kul-button\",{\"kulData\":[1040],\"kulDisabled\":[1540,\"kul-disabled\"],\"kulIcon\":[1537,\"kul-icon\"],\"kulIconOff\":[1537,\"kul-icon-off\"],\"kulLabel\":[1537,\"kul-label\"],\"kulRipple\":[1540,\"kul-ripple\"],\"kulShowSpinner\":[1540,\"kul-show-spinner\"],\"kulStyle\":[1537,\"kul-style\"],\"kulStyling\":[1537,\"kul-styling\"],\"kulToggable\":[1540,\"kul-toggable\"],\"kulTrailingIcon\":[1540,\"kul-trailing-icon\"],\"kulType\":[1537,\"kul-type\"],\"kulValue\":[4,\"kul-value\"],\"debugInfo\":[32],\"value\":[32],\"getDebugInfo\":[64],\"getProps\":[64],\"getValue\":[64],\"refresh\":[64],\"setValue\":[64],\"unmount\":[64]}],[1,\"kul-list\",{\"kulData\":[1040],\"kulEmptyLabel\":[1,\"kul-empty-label\"],\"kulEnableDeletions\":[4,\"kul-enable-deletions\"],\"kulNavigation\":[4,\"kul-navigation\"],\"kulRipple\":[1540,\"kul-ripple\"],\"kulSelectable\":[4,\"kul-selectable\"],\"kulStyle\":[1025,\"kul-style\"],\"debugInfo\":[32],\"focused\":[32],\"selected\":[32],\"focusNext\":[64],\"focusPrevious\":[64],\"getDebugInfo\":[64],\"getProps\":[64],\"getSelected\":[64],\"refresh\":[64],\"selectNode\":[64],\"unmount\":[64]},[[0,\"keydown\",\"listenKeydown\"]]],[1,\"kul-badge\",{\"kulImageProps\":[1040],\"kulLabel\":[1025,\"kul-label\"],\"kulStyle\":[1537,\"kul-style\"],\"debugInfo\":[32],\"getDebugInfo\":[64],\"getProps\":[64],\"refresh\":[64],\"unmount\":[64]}],[1,\"kul-image\",{\"kulBadgeProps\":[1040],\"kulColor\":[1537,\"kul-color\"],\"kulShowSpinner\":[1540,\"kul-show-spinner\"],\"kulSizeX\":[1537,\"kul-size-x\"],\"kulSizeY\":[1537,\"kul-size-y\"],\"kulStyle\":[1537,\"kul-style\"],\"kulValue\":[1537,\"kul-value\"],\"debugInfo\":[32],\"getDebugInfo\":[64],\"getProps\":[64],\"refresh\":[64],\"unmount\":[64]}]]],[\"kul-chat_3\",[[1,\"kul-chat\",{\"kulContextWindow\":[1026,\"kul-context-window\"],\"kulEndpointUrl\":[1025,\"kul-endpoint-url\"],\"kulLayout\":[1537,\"kul-layout\"],\"kulMaxTokens\":[1026,\"kul-max-tokens\"],\"kulPollingInterval\":[2,\"kul-polling-interval\"],\"kulSeed\":[1026,\"kul-seed\"],\"kulStyle\":[1537,\"kul-style\"],\"kulSystem\":[1025,\"kul-system\"],\"kulTemperature\":[1026,\"kul-temperature\"],\"kulValue\":[1040],\"debugInfo\":[32],\"history\":[32],\"status\":[32],\"toolbarMessage\":[32],\"view\":[32],\"getDebugInfo\":[64],\"getHistory\":[64],\"getLastMessage\":[64],\"getProps\":[64],\"refresh\":[64],\"setHistory\":[64],\"unmount\":[64]},[[0,\"keydown\",\"listenKeydown\"]],{\"kulSystem\":[\"updateTokensCount\"]}],[1,\"kul-tabbar\",{\"kulData\":[16],\"kulRipple\":[1540,\"kul-ripple\"],\"kulStyle\":[1537,\"kul-style\"],\"kulValue\":[520,\"kul-value\"],\"debugInfo\":[32],\"value\":[32],\"getDebugInfo\":[64],\"getProps\":[64],\"getValue\":[64],\"refresh\":[64],\"setValue\":[64],\"unmount\":[64]}],[1,\"kul-chip\",{\"kulData\":[1040],\"kulRipple\":[1540,\"kul-ripple\"],\"kulStyle\":[1025,\"kul-style\"],\"kulStyling\":[1025,\"kul-styling\"],\"debugInfo\":[32],\"expandedNodes\":[32],\"hiddenNodes\":[32],\"selectedNodes\":[32],\"getDebugInfo\":[64],\"getProps\":[64],\"getSelectedNodes\":[64],\"refresh\":[64],\"setSelectedNodes\":[64],\"unmount\":[64]}]]]]"), options);
 });
 
 //# sourceMappingURL=ketchup-lite.js.map
