+++ web/deploy/js/ketchup-lite/kul-accordion.entry.js
@@ -1,6 +1,6 @@
 import { r as registerInstance, d as createEvent, g as getElement, f as forceUpdate, a as getAssetPath, h, H as Host } from './index-4d533537.js';
 import { a as KulDataCyAttributes, K as KUL_WRAPPER_ID, c as KUL_STYLE_ID } from './GenericVariables-f3380974.js';
-import { k as kulManagerInstance } from './kul-manager-8d12091b.js';
+import { k as kulManagerInstance } from './kul-manager-5dda66e6.js';
 import { g as getProps } from './componentUtils-a994b230.js';
 
 /*-------------------------------------------------*/
