+++ web/deploy/js/ketchup-lite/kul-badge_4.entry.js
@@ -1,7 +1,121 @@
 import { r as registerInstance, d as createEvent, g as getElement, f as forceUpdate, h, H as Host, F as Fragment, a as getAssetPath } from './index-4d533537.js';
-import { b as KUL_DROPDOWN_CLASS_VISIBLE, d as KUL_DROPDOWN_CLASS, a as KulDataCyAttributes, K as KUL_WRAPPER_ID, c as KUL_STYLE_ID } from './GenericVariables-f3380974.js';
-import { k as kulManagerInstance, d as KulDynamicPositionPlacement, b as KulLanguageGeneric } from './kul-manager-8d12091b.js';
+import { k as kulManagerInstance, K as KulThemeColorValues, d as KulDynamicPositionPlacement, b as KulLanguageGeneric } from './kul-manager-5dda66e6.js';
 import { g as getProps } from './componentUtils-a994b230.js';
+import { K as KUL_WRAPPER_ID, c as KUL_STYLE_ID, b as KUL_DROPDOWN_CLASS_VISIBLE, d as KUL_DROPDOWN_CLASS, a as KulDataCyAttributes, C as CSS_VAR_PREFIX } from './GenericVariables-f3380974.js';
+
+/*-------------------------------------------------*/
+/*                    P r o p s                    */
+/*-------------------------------------------------*/
+var KulBadgeProps;
+(function (KulBadgeProps) {
+    KulBadgeProps["kulImageProps"] = "The props of the image displayed inside the badge.";
+    KulBadgeProps["kulLabel"] = "The text displayed inside the badge.";
+    KulBadgeProps["kulStyle"] = "Custom style of the component.";
+})(KulBadgeProps || (KulBadgeProps = {}));
+
+const kulBadgeCss = ".ripple-surface{cursor:pointer;height:100%;left:0;overflow:hidden;position:absolute;top:0;width:100%}.ripple{animation:ripple 0.675s ease-out;border-radius:50%;pointer-events:none;position:absolute;transform:scale(0)}@keyframes ripple{to{opacity:0;transform:scale(4)}}::-webkit-scrollbar{width:9px}::-webkit-scrollbar-thumb{background-color:var(--kul-primary-color);-webkit-transition:background-color 0.2s ease-in-out;transition:background-color 0.2s ease-in-out}::-webkit-scrollbar-track{background-color:var(--kul-background-color)}@keyframes fade-in-block{0%{display:none}1%{display:block;opacity:0}100%{display:block;opacity:1}}@keyframes fade-in-flex{0%{display:none}1%{display:flex;opacity:0}100%{display:flex;opacity:1}}:host{--kul_badge_border_radius:var(--kul-badge-border-radius, 30px);--kul_badge_font_family:var(--kul-badge-font-family, var(--kul-font-family));--kul_badge_font_size:var(--kul-badge-font-size, var(--kul-font-size));--kul_badge_min_size:var(--kul-badge-min-size, 1.5em);--kul_badge_padding:var(--kul-badge-padding, 0.25em);--kul_badge_primary_color:var(\n    --kul-badge-primary-color,\n    var(--kul-primary-color)\n  );--kul_badge_text_on_primary_color:var(\n    --kul-badge-text-on-primary-color,\n    var(--kul-text-on-primary-color)\n  );display:block;font-family:var(--kul_badge_font_family);font-size:var(--kul_badge_font_size);position:absolute;top:0;left:0;transform:translate(-50%, -50%)}#kul-component{background-color:var(--kul_badge_primary_color);border-radius:var(--kul_badge_border_radius);color:var(--kul_badge_text_on_primary_color);font-size:0.875em;min-height:var(--kul_badge_min_size);min-width:var(--kul_badge_min_size);padding:var(--kul_badge_padding);text-align:center}kul-image{left:50%;position:absolute;top:50%;transform:translate(-50%, -50%)}:host(.kul-top-right){bottom:unset;left:unset;right:0;top:0;transform:translate(50%, -50%)}:host(.kul-bottom-right){bottom:0;left:unset;right:0;top:unset;transform:translate(50%, 50%)}:host(.kul-bottom-left){bottom:0;left:0;right:unset;top:unset;transform:translate(-50%, 50%)}:host(.kul-danger){--kul-badge-primary-color:var(--kul-danger-color);--kul-badge-text-on-primary-color:white}:host(.kul-info){--kul-badge-primary-color:var(--kul-info-color);--kul-badge-text-on-primary-color:white}:host(.kul-secondary){--kul-badge-primary-color:var(--kul-secondary-color);--kul-badge-text-on-primary-color:var(--kul-text-on-secondary-color)}:host(.kul-success){--kul-badge-primary-color:var(--kul-success-color);--kul-badge-text-on-primary-color:white}:host(.kul-warning){--kul-badge-primary-color:var(--kul-warning-color);--kul-badge-text-on-primary-color:white}";
+const KulBadgeStyle0 = kulBadgeCss;
+
+const KulBadge = class {
+    constructor(hostRef) {
+        registerInstance(this, hostRef);
+        this.kulEvent = createEvent(this, "kul-badge-event", 6);
+        this.debugInfo = {
+            endTime: 0,
+            renderCount: 0,
+            renderEnd: 0,
+            renderStart: 0,
+            startTime: performance.now(),
+        };
+        this.kulImageProps = null;
+        this.kulLabel = '';
+        this.kulStyle = '';
+    }
+    get rootElement() { return getElement(this); }
+    /*-------------------------------------------------*/
+    /*       I n t e r n a l   V a r i a b l e s       */
+    /*-------------------------------------------------*/
+    #kulManager = kulManagerInstance();
+    /*-------------------------------------------------*/
+    /*                   E v e n t s                   */
+    /*-------------------------------------------------*/
+    /**
+     * Describes event emitted.
+     */
+    kulEvent;
+    onKulEvent(e, eventType) {
+        this.kulEvent.emit({
+            comp: this,
+            eventType,
+            id: this.rootElement.id,
+            originalEvent: e,
+        });
+    }
+    /*-------------------------------------------------*/
+    /*           P u b l i c   M e t h o d s           */
+    /*-------------------------------------------------*/
+    /**
+     * Fetches debug information of the component's current state.
+     * @returns {Promise<KulDebugLifecycleInfo>} A promise that resolves with the debug information object.
+     */
+    async getDebugInfo() {
+        return this.debugInfo;
+    }
+    /**
+     * Used to retrieve component's props values.
+     * @param {boolean} descriptions - When provided and true, the result will be the list of props with their description.
+     * @returns {Promise<GenericObject>} List of props as object, each key will be a prop.
+     */
+    async getProps(descriptions) {
+        return getProps(this, KulBadgeProps, descriptions);
+    }
+    /**
+     * This method is used to trigger a new render of the component.
+     */
+    async refresh() {
+        forceUpdate(this);
+    }
+    /**
+     * Initiates the unmount sequence, which removes the component from the DOM after a delay.
+     * @param {number} ms - Number of milliseconds
+     */
+    async unmount(ms = 0) {
+        setTimeout(() => {
+            this.onKulEvent(new CustomEvent('unmount'), 'unmount');
+            this.rootElement.remove();
+        }, ms);
+    }
+    /*-------------------------------------------------*/
+    /*          L i f e c y c l e   H o o k s          */
+    /*-------------------------------------------------*/
+    componentWillLoad() {
+        this.#kulManager.theme.register(this);
+    }
+    componentDidLoad() {
+        this.onKulEvent(new CustomEvent('ready'), 'ready');
+        this.#kulManager.debug.updateDebugInfo(this, 'did-load');
+    }
+    componentWillRender() {
+        this.#kulManager.debug.updateDebugInfo(this, 'will-render');
+    }
+    componentDidRender() {
+        this.#kulManager.debug.updateDebugInfo(this, 'did-render');
+    }
+    render() {
+        let imageEl = null;
+        if (!this.kulLabel && this.kulImageProps) {
+            if (!this.kulImageProps.kulColor) {
+                this.kulImageProps.kulColor = `var(${KulThemeColorValues.TEXT_ON_PRIMARY})`;
+            }
+            imageEl = h("kul-image", { key: '7b791716faeb2d64a1efc25ac587dc7b4188fef7', ...this.kulImageProps });
+        }
+        return (h(Host, { key: '2b083832a5469320c830a34a4c42633e4e9650b1' }, this.kulStyle ? (h("style", { id: KUL_STYLE_ID }, this.#kulManager.theme.setKulStyle(this))) : undefined, h("div", { key: '6f22d25c8907dfdf7f225c079fbd00dc4cb25682', id: KUL_WRAPPER_ID, onClick: (e) => this.onKulEvent(e, 'click') }, this.kulLabel, imageEl)));
+    }
+    disconnectedCallback() {
+        this.#kulManager.theme.unregister(this);
+    }
+};
+KulBadge.style = KulBadgeStyle0;
 
 /*-------------------------------------------------*/
 /*                    P r o p s                    */
@@ -396,6 +510,190 @@ KulButton.style = KulButtonStyle0;
 /*-------------------------------------------------*/
 /*                    P r o p s                    */
 /*-------------------------------------------------*/
+var KulImageProps;
+(function (KulImageProps) {
+    KulImageProps["kulBadgeProps"] = "Sets the props to show a badge.";
+    KulImageProps["kulColor"] = "The color of the icon, defaults to the CSS variable --kul-icon-color.";
+    KulImageProps["kulShowSpinner"] = "When set to true, a spinner will be displayed until the image finished loading. Not compatible with SVGs.";
+    KulImageProps["kulSizeX"] = "The width of the icon, defaults to 100%. Accepts any valid CSS format (px, %, vh, etc.).";
+    KulImageProps["kulSizeY"] = "The height of the icon, defaults to 100%. Accepts any valid CSS format (px, %, vh, etc.).";
+    KulImageProps["kulStyle"] = "Custom style of the component.";
+    KulImageProps["kulValue"] = "Defines the source URL of the image. This property is used to set the image resource that the component should display.";
+})(KulImageProps || (KulImageProps = {}));
+
+const kulImageCss = ".ripple-surface{cursor:pointer;height:100%;left:0;overflow:hidden;position:absolute;top:0;width:100%}.ripple{animation:ripple 0.675s ease-out;border-radius:50%;pointer-events:none;position:absolute;transform:scale(0)}@keyframes ripple{to{opacity:0;transform:scale(4)}}::-webkit-scrollbar{width:9px}::-webkit-scrollbar-thumb{background-color:var(--kul-primary-color);-webkit-transition:background-color 0.2s ease-in-out;transition:background-color 0.2s ease-in-out}::-webkit-scrollbar-track{background-color:var(--kul-background-color)}@keyframes fade-in-block{0%{display:none}1%{display:block;opacity:0}100%{display:block;opacity:1}}@keyframes fade-in-flex{0%{display:none}1%{display:flex;opacity:0}100%{display:flex;opacity:1}}:host{display:block;height:var(--kul_image_height, 100%);margin:auto;min-height:var(--kul_image_height, 100%);min-width:var(--kul_image_width, 100%);position:relative;transition:color 0.2s ease;width:var(--kul_image_width, 100%)}#kul-component{height:100%;margin:auto;position:relative;transition:color 0.2s ease;width:100%}.image{display:block;height:var(--kul_image_height, 100%);margin:auto;position:relative;width:var(--kul_image_width, 100%)}.image__icon{height:var(--kul_image_height, 100%);margin:auto;width:var(--kul_image_width, 100%)}img{display:block;height:var(--kul_image_height, 100%);margin:auto;width:var(--kul_image_width, 100%)}.spinner{height:32px;left:calc(50% - 16px);position:absolute;top:calc(50% - 16px);width:32px}:host(.kul-fit) img{max-width:max-content;object-fit:contain}:host(.kul-cover) img{object-fit:cover}";
+const KulImageStyle0 = kulImageCss;
+
+const KulImage = class {
+    constructor(hostRef) {
+        registerInstance(this, hostRef);
+        this.kulEvent = createEvent(this, "kul-image-event", 6);
+        this.debugInfo = {
+            endTime: 0,
+            renderCount: 0,
+            renderEnd: 0,
+            renderStart: 0,
+            startTime: performance.now(),
+        };
+        this.kulBadgeProps = null;
+        this.kulColor = `var(${KulThemeColorValues.ICON})`;
+        this.kulShowSpinner = false;
+        this.kulSizeX = '100%';
+        this.kulSizeY = '100%';
+        this.kulStyle = '';
+        this.kulValue = '';
+    }
+    get rootElement() { return getElement(this); }
+    /*-------------------------------------------------*/
+    /*        I n t e r n a l   V a r i a b l e s      */
+    /*-------------------------------------------------*/
+    #kulManager = kulManagerInstance();
+    /*-------------------------------------------------*/
+    /*                   E v e n t s                   */
+    /*-------------------------------------------------*/
+    /**
+     * Describes event emitted.
+     */
+    kulEvent;
+    onKulEvent(e, eventType) {
+        this.kulEvent.emit({
+            comp: this,
+            id: this.rootElement.id,
+            originalEvent: e,
+            eventType,
+        });
+    }
+    /*-------------------------------------------------*/
+    /*           P u b l i c   M e t h o d s           */
+    /*-------------------------------------------------*/
+    /**
+     * Fetches debug information of the component's current state.
+     * @returns {Promise<KulDebugLifecycleInfo>} A promise that resolves with the debug information object.
+     */
+    async getDebugInfo() {
+        return this.debugInfo;
+    }
+    /**
+     * Used to retrieve component's props values.
+     * @param {boolean} descriptions - When provided and true, the result will be the list of props with their description.
+     * @returns {Promise<GenericObject>} List of props as object, each key will be a prop.
+     */
+    async getProps(descriptions) {
+        return getProps(this, KulImageProps, descriptions);
+    }
+    /**
+     * This method is used to trigger a new render of the component.
+     */
+    async refresh() {
+        forceUpdate(this);
+    }
+    /**
+     * Initiates the unmount sequence, which removes the component from the DOM after a delay.
+     * @param {number} ms - Number of milliseconds
+     */
+    async unmount(ms = 0) {
+        setTimeout(() => {
+            this.onKulEvent(new CustomEvent('unmount'), 'unmount');
+            this.rootElement.remove();
+        }, ms);
+    }
+    /*-------------------------------------------------*/
+    /*           P r i v a t e   M e t h o d s         */
+    /*-------------------------------------------------*/
+    createIcon() {
+        const classObj = {
+            image__icon: true,
+        };
+        const style = {
+            background: this.kulColor
+                ? this.kulColor
+                : `var(${KulThemeColorValues.ICON})`,
+        };
+        if (this.kulValue.indexOf(CSS_VAR_PREFIX) > -1) {
+            const themeIcon = this.kulValue.replace('--', '');
+            classObj['kul-icon'] = true;
+            classObj[themeIcon] = true;
+            const icon = this.#kulManager.theme.list[this.#kulManager.theme.name].icons[this.kulValue];
+            const path = getAssetPath(`./assets/svg/${icon}.svg`);
+            style.mask = `url('${path}') no-repeat center`;
+            style.webkitMask = `url('${path}') no-repeat center`;
+        }
+        else {
+            const path = getAssetPath(`./assets/svg/${this.kulValue}.svg`);
+            style.mask = `url('${path}') no-repeat center`;
+            style.webkitMask = `url('${path}') no-repeat center`;
+        }
+        return h("div", { class: classObj, style: style });
+    }
+    createImage() {
+        return (h("img", { onLoad: (e) => {
+                this.onKulEvent(e, 'load');
+            }, src: this.kulValue }));
+    }
+    isResourceUrl() {
+        return !!(this.kulValue &&
+            (this.kulValue.indexOf('.') > -1 ||
+                this.kulValue.indexOf('/') > -1 ||
+                this.kulValue.indexOf('\\') > -1));
+    }
+    /*-------------------------------------------------*/
+    /*          L i f e c y c l e   H o o k s          */
+    /*-------------------------------------------------*/
+    componentWillLoad() {
+        this.#kulManager.theme.register(this);
+    }
+    componentDidLoad() {
+        this.onKulEvent(new CustomEvent('ready'), 'ready');
+        this.#kulManager.debug.updateDebugInfo(this, 'did-load');
+    }
+    componentWillRender() {
+        this.#kulManager.debug.updateDebugInfo(this, 'will-render');
+    }
+    componentDidRender() {
+        this.#kulManager.debug.updateDebugInfo(this, 'did-render');
+    }
+    render() {
+        if (!this.kulValue) {
+            this.#kulManager.debug.logs.new(this, 'Empty image.');
+            return;
+        }
+        let el;
+        let feedback;
+        const isUrl = this.isResourceUrl();
+        let spinnerLayout;
+        let style;
+        if (isUrl) {
+            style = {
+                '--kul_image_height': this.kulSizeY ? this.kulSizeY : 'auto',
+                '--kul_image_width': this.kulSizeX ? this.kulSizeX : '100%',
+            };
+            el = this.createImage();
+        }
+        else {
+            style = {
+                '--kul_image_height': this.kulSizeY ? this.kulSizeY : '100%',
+                '--kul_image_width': this.kulSizeX ? this.kulSizeX : '100%',
+            };
+            el = this.createIcon();
+        }
+        if (this.kulShowSpinner && isUrl) {
+            spinnerLayout = 14;
+            feedback = (h("div", { class: "spinner", title: "Image not loaded yet..." }, h("kul-spinner", { kulActive: true, kulDimensions: "3px", kulLayout: spinnerLayout })));
+        }
+        return (h(Host, { style: style }, this.kulStyle ? (h("style", { id: KUL_STYLE_ID }, this.#kulManager.theme.setKulStyle(this))) : undefined, feedback, h("div", { id: KUL_WRAPPER_ID }, h("div", { class: "image", onClick: (e) => {
+                this.onKulEvent(e, 'click');
+            } }, el, this.kulBadgeProps ? (h("kul-badge", { ...this.kulBadgeProps })) : undefined))));
+    }
+    disconnectedCallback() {
+        this.#kulManager.theme.unregister(this);
+    }
+    static get assetsDirs() { return ["assets/svg"]; }
+};
+KulImage.style = KulImageStyle0;
+
+/*-------------------------------------------------*/
+/*                    P r o p s                    */
+/*-------------------------------------------------*/
 var KulListProps;
 (function (KulListProps) {
     KulListProps["kulData"] = "The actual data of the list.";
@@ -681,6 +979,6 @@ const KulList = class {
 };
 KulList.style = KulListStyle0;
 
-export { KulButton as kul_button, KulList as kul_list };
+export { KulBadge as kul_badge, KulButton as kul_button, KulImage as kul_image, KulList as kul_list };
 
-//# sourceMappingURL=kul-button_2.entry.js.map
+//# sourceMappingURL=kul-badge_4.entry.js.map
