+++ web/deploy/js/ketchup-lite/kul-manager-5dda66e6.js
@@ -204,17 +204,23 @@ const decorateSpreader = (toSpread, props) => {
         if (toSpread['value'] && !toSpread['kulValue']) {
             toSpread['kulValue'] = toSpread['value'];
         }
+        delete toSpread['htmlProps'];
         delete toSpread['shape'];
         delete toSpread['value'];
     };
     if (props.htmlProps) {
         for (const key in props.htmlProps) {
             const prop = props.htmlProps[key];
-            if (prop === 'className') {
+            if (key === 'className') {
                 toSpread['class'] = prop;
             }
             else {
                 toSpread[key] = prop;
+            }
+            if (key === 'dataset') {
+                for (const k in prop) {
+                    toSpread[`data-${k}`] = prop[k];
+                }
             }
         }
     }
@@ -6846,4 +6852,4 @@ function kulManagerInstance() {
 
 export { KulThemeColorValues as K, KulLanguageSearch as a, KulLanguageGeneric as b, commonjsGlobal as c, KulDynamicPositionPlacement as d, kulManagerInstance as k };
 
-//# sourceMappingURL=kul-manager-8d12091b.js.map
+//# sourceMappingURL=kul-manager-5dda66e6.js.map
